feat/build: cross platform test target
Simple test target that will use the container on Linux or just run the
formatting and testing with Cargo on macOS or Windows. The reason why
the container needs to use /bin/bash -c, is because, if you attempt to
run these commands directly without that, any commands after the first
'&&' will run on the *host* rather than the container.

This also installs clippy and rustfmt in the container and runs clippy
during the build process.